requirements.txt
requests>=2.31.0
websocket-client>=1.6.3
pycryptodome>=3.15.0
eth-account~=0.13.3
pathlib~=1.0.1